# docs/openapi.yaml
components:
  schemas:
...
    CoreBuildResponse:
      additionalProperties: false
      properties:
        created_at:
          type: string
        id:
          format: int64
          type: integer
        parent_build_id:
          format: int64
          type: integer
        state:
          type: string
        upload_id:
          format: int64
          type: integer
        user_version:
          type: string
      required:
        - id
        - user_version
        - state
        - created_at
      type: object
    CoreUploadResponse:
      additionalProperties: false
      properties:
        display_name:
          type: string
        filename:
          type: string
        id:
          format: int64
          type: integer
        platforms:
          description: JSON array encoded as a string.
          type: string
        size:
          format: int64
          type: integer
        storage:
          type: string
        type:
          type: string
      required:
        - id
        - filename
        - display_name
        - size
        - storage
        - type
        - platforms
      type: object
...
    GameResponse:
      additionalProperties: false
      properties:
        classification:
          type: string
        id:
          format: int64
          type: integer
        short_text:
          type: string
        title:
          type: string
        type:
          type: string
        url:
          type: string
        user:
          $ref: "#/components/schemas/PublicUserResponse"
      required:
        - id
        - title
        - short_text
        - type
        - classification
        - url
        - user
      type: object
...
    PublicUserResponse:
      additionalProperties: false
      properties:
        display_name:
          type: string
        id:
          format: int64
          type: integer
        username:
          type: string
      required:
        - id
        - username
        - display_name
      type: object
    UpgradePathBody:
      additionalProperties: false
      properties:
        builds:
          items:
            $ref: "#/components/schemas/WharfBuildResponse"
          type:
            - array
            - "null"
      required:
        - builds
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      description: Raw API key, Bearer token, or access_token token.
      in: header
      name: Authorization
      type: apiKey
info:
  title: DenKit Stash API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /builds/{buildId}/download/{type}/{subType}:
    get:
      operationId: download-build-artifact
      parameters:
        - in: path
          name: buildId
          required: true
          schema:
            format: int64
            type: integer
        - in: path
          name: type
          required: true
          schema:
            type: string
        - in: path
          name: subType
          required: true
          schema:
            type: string
      responses:
        "307":
          description: Temporary Redirect
          headers:
            Location:
              schema:
                format: uri
                type: string
        "400":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Not Found
      security:
        - ApiKeyAuth: []
      summary: Redirect to a build artifact by type and subtype
      tags:
        - Downloads
    head:
      operationId: head-build-artifact
      parameters:
        - in: path
          name: buildId
          required: true
          schema:
            format: int64
            type: integer
        - in: path
          name: type
          required: true
          schema:
            type: string
        - in: path
          name: subType
          required: true
          schema:
            type: string
      responses:
        "307":
          description: Temporary Redirect
          headers:
            Location:
              schema:
                format: uri
                type: string
        "401":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Unauthorized
      security:
        - ApiKeyAuth: []
      summary: Redirect to a build artifact by type and subtype
      tags:
        - Downloads
  /builds/{id}:
    get:
      operationId: get-build
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: int64
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  build:
                    $ref: "#/components/schemas/CoreBuildResponse"
                required:
                  - build
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Not Found
      security:
        - ApiKeyAuth: []
        - {}
      summary: Get a build by ID
      tags:
        - Core
  /builds/{installedBuildId}/upgrade-paths/{targetBuildId}:
    get:
      operationId: get-upgrade-path
      parameters:
        - in: path
          name: installedBuildId
          required: true
          schema:
            format: int64
            type: integer
        - in: path
          name: targetBuildId
          required: true
          schema:
            format: int64
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  upgradePath:
                    $ref: "#/components/schemas/UpgradePathBody"
                required:
                  - upgradePath
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Not Found
      security:
        - ApiKeyAuth: []
      summary: Resolve patch/signature upgrade path between builds
      tags:
        - Downloads
  /games/{id}:
    get:
      operationId: get-game
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: int64
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  game:
                    $ref: "#/components/schemas/GameResponse"
                required:
                  - game
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Not Found
      security:
        - ApiKeyAuth: []
        - {}
      summary: Get a game by ID
      tags:
        - Core
  /games/{id}/download-sessions:
    post:
      operationId: create-download-session
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: int64
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  uuid:
                    format: uuid
                    type: string
                required:
                  - uuid
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Not Found
      security:
        - ApiKeyAuth: []
        - {}
      summary: Create a download session UUID
      tags:
        - Core
  /games/{id}/uploads:
    get:
      operationId: list-game-uploads
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: int64
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  uploads:
                    items:
                      $ref: "#/components/schemas/CoreUploadResponse"
                    type:
                      - array
                      - "null"
                required:
                  - uploads
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Not Found
      security:
        - ApiKeyAuth: []
        - {}
      summary: List uploads for a game
      tags:
        - Core
  /profile:
    get:
      operationId: get-profile
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  user:
                    $ref: "#/components/schemas/PublicUserResponse"
                required:
                  - user
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Unauthorized
      security:
        - ApiKeyAuth: []
      summary: Get the authenticated user profile
      tags:
        - Core
  /profile/games:
    get:
      operationId: list-profile-games
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  games:
                    items:
                      $ref: "#/components/schemas/ProfileGameResponse"
                    type:
                      - array
                      - "null"
                required:
                  - games
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Unauthorized
      security:
        - ApiKeyAuth: []
      summary: List games owned by the authenticated user
      tags:
        - Core
  /uploads/{id}:
    get:
      operationId: get-upload
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: int64
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  upload:
                    $ref: "#/components/schemas/CoreUploadResponse"
                required:
                  - upload
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Not Found
      security:
        - ApiKeyAuth: []
        - {}
      summary: Get an upload by ID
      tags:
        - Core
  /uploads/{id}/builds:
    get:
      operationId: list-upload-builds
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: int64
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  builds:
                    items:
                      $ref: "#/components/schemas/CoreBuildResponse"
                    type:
                      - array
                      - "null"
                required:
                  - builds
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Not Found
      security:
        - ApiKeyAuth: []
        - {}
      summary: List builds for an upload
      tags:
        - Core
...
  /wharf/status:
    get:
      operationId: get-wharf-status
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  ok:
                    examples:
                      - true
                    type: boolean
                required:
                  - ok
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Unauthorized
      security:
        - ApiKeyAuth: []
      summary: Check Wharf-compatible API status
      tags:
        - Wharf
...